🚚(frontend) rename ContractFilters component
ContractFilters is too generic as a components name.
